+++ ormconfig.json
@@ -6,8 +6,8 @@
   "password": "docker",
   "database": "shorten-link",
   "entities": ["./src/entities/*.ts"],
-  "migrations": ["./src/database/migrations/*.ts"],
+  "migrations": ["./src/shared/infra/database/migrations/*.ts"],
   "cli": {
-    "migrationsDir": "./src/database/migrations"
+    "migrationsDir": "./src/shared/infra/database/migrations"
   }
 }
