+++ package.json
@@ -8,7 +8,7 @@
   "license": "MIT",
   "private": false,
   "scripts": {
-    "start:dev": "ts-node-dev --transpile-only --ignore-watch node_modules src/server.ts",
+    "start:dev": "ts-node-dev --transpile-only --ignore-watch node_modules src/shared/infra/http/server.ts",
     "test": "jest --forceExit --silent --runInBand --noStackTrace",
     "test:watch": "jest --watch",
     "test:cov": "jest --coverage",
