--- src/container/index.ts
@@ -1,8 +1,0 @@
-import { container } from 'tsyringe';
-
-import { UrlRepository } from '../database/repositories/implementations/UrlRepository';
-import { IUrlRepository } from '../database/repositories/IUrlRepository';
-
-container.registerSingleton<IUrlRepository>('UrlRepository', UrlRepository);
-
-export default container;
