+++ src/shared/infra/http/server.ts
@@ -3,11 +3,11 @@ import { ApolloServer } from 'apollo-server-express';
 import { ApolloServerPluginLandingPageGraphQLPlayground } from 'apollo-server-core';
 import express from 'express';
 import { buildSchema } from 'type-graphql';
-import { UrlResolver } from './resolvers/Url';
-import { enviroments } from './configs/enviroments';
+import { UrlResolver } from '../../../modules/url/resolvers/Url';
+import { enviroments } from '../../configs/enviroments';
 
-import './container';
-import createConnection from './database';
+import '../../container';
+import createConnection from '../database';
 
 const init = async () => {
   createConnection();
